Moved shared testing code to a new file.
